% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dailySummaryPlot.R
\name{dailySummaryPlot}
\alias{dailySummaryPlot}
\title{Generate summary plots from user data}
\usage{
dailySummaryPlot(dataTable, outcome = "calories", alsoMeal = FALSE,
  recommended = NULL, title = paste("daily intake of", outcome))
}
\arguments{
\item{dataTable}{a table of food or exercise elements from \code{\link{scrappal}}}

\item{outcome}{a column name of dataTable to put on y axis}

\item{alsoMeal}{should we also break it down by meal}
}
\value{
a plot to be used in a shiny app
}
\description{
Generate summary plots from user data
}
\examples{
\dontrun{

username = 'scrappal'
fromDate = as.Date('2015-08-01','\%Y-\%m-\%d')
toDate =   as.Date('2015-08-10','\%Y-\%m-\%d')

z<-scrappal(
		username = username,
		fromDate = fromDate,
		toDate = toDate,
		includeDailyCal = TRUE
		)

dailySummaryPlot(dataTable=z$food, outcome='calories', alsoMeal=FALSE)
dailySummaryPlot(dataTable=z$food, outcome='fat', alsoMeal=TRUE)
dailySummaryPlot(dataTable=z$exercise, outcome='calories', alsoMeal=FALSE, title='calories burned')
dailySummaryPlot(dataTable=z$exercise, outcome='minutes', alsoMeal=FALSE, title='exercise per day (minutes)')

}
}

