% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rvesting.R
\name{scrappal}
\alias{scrappal}
\title{Scrape data from my fitness pal}
\usage{
scrappal(username = "bcaffo", fromDate, toDate,
  verbose = getOption("verbose"), includeDailyCal = TRUE)
}
\arguments{
\item{username}{for person of interest}

\item{fromDate}{date to start gathering R data. Should have class "Date"}

\item{toDate}{date to stop gathering R data. Should have class "Date"}

\item{verbose}{whether to print progress}
}
\value{
A list containing tables of: all exercises, all food items eaten, and the calories burned and consumed on every day.
}
\description{
Scrape data from my fitness pal
}
\examples{
\dontrun{
username = 'funchords'
fromDate = as.Date('2015-08-01','\%Y-\%m-\%d')
toDate =   as.Date('2015-08-10','\%Y-\%m-\%d')

z<-scrappal(
		username = username,
		fromDate = fromDate,
		toDate = toDate
		)

### Show food labels separately, because they're large
z$food[1:10,1]
### preview everything but the label
z$food[,-1]

z$exercise
z$dailyCal

}
}

